Add Google Analytics.
I have a lot of mixed feelings about doing this.

On the one hand, I am an advocate of privacy.

On the other hand, I need to be able to gauge interest in this project,
and for different facets of it, so I can appropriately allocate my
scarce time and effort.

If any concerned readers can offer a better solution for analytics
that will better protect privacy of users, please file an issue ticket
or pull request and we can have an open discussion about it!