Change TREG semantics to use 2-element array for empty state GET.
This helps fix it to follow the rule that missing CRDT values should
be treated as being in the "bottom state" - empty rather than missing.
It helps with consistency and making the result easier to deal with
on the client side because the result will always have the same format.