Hacky solution to issue #10, needing a clean followup later.
This commit uses a brittle hack to remove the PDAT Msg headers
from replicated data being stored to disk. This hack will break
if/when the clustering/disk protocol changes.

However, a clean solution probably requires a change to the protocol.
Specifically, we'd want to encode multiple RESP objects in a single Frame,
so that the first object is the Msg header and the second is the data itself.

I'll withold a clean solution for now, until we're ready to make changes to the protocol.

Resolves #10.